requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.1.1
django-jquery==3.1.0
django-widget-tweaks==1.4.8
future==0.18.2
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
viberbot==1.0.11
